requirements.txt
attrs==19.3.0
beautifulsoup4==4.8.2
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
flaky==3.6.1
future==0.18.2
geojson==2.5.0
idna==2.8
importlib-metadata==1.4.0
more-itertools==8.1.0
mysqlclient==1.4.6
packaging==20.1
pluggy==0.13.1
psutil==5.6.7
py==1.8.1
pycparser==2.19
pyowm==2.10.0
pyparsing==2.4.6
pytest==5.3.4
python-telegram-bot==12.3.0
requests==2.22.0
six==1.14.0
soupsieve==1.9.5
tornado==6.0.3
urllib3==1.25.8
wcwidth==0.1.8
wikipedia==1.4.0
yandex-translater==6.0
zipp==2.1.0